feat(tao): unify Windows a11y on AccessKit with enterprise UIA probes
Replace hand-rolled nucleus_tao_a11y COM projector with AccessKit on Windows,
share wire/decode path with Linux, and add golden tree + keyboard/event CI probes.
