+++ decorated-window-tao/src/main/native/windows/build.bat
@@ -20,7 +20,6 @@ set "NATIVE_DIR=%SCRIPT_DIR%.."
 set "DECO_SRC=%SCRIPT_DIR%nucleus_tao_windows_deco.c"
 
 set "GL_SRC=%SCRIPT_DIR%nucleus_tao_gl.c"
-set "A11Y_SRC=%SCRIPT_DIR%nucleus_tao_a11y.c"
 set "DND_SRC=%SCRIPT_DIR%nucleus_tao_dnd.c"
 set "NV_SRC=%SCRIPT_DIR%nucleus_tao_windows_native_view.c"
 set "OVERLAY_SRC=%SCRIPT_DIR%nucleus_tao_windows_overlay.c"
@@ -152,20 +151,6 @@ if errorlevel 1 (
     exit /b 1
 )
 
-REM nucleus_tao_a11y.dll — UI Automation provider. Needs the C runtime
-REM (HeapAlloc-friendly libcmt) because uiautomationcore.lib pulls in C++
-REM statics that require __security_check_cookie / floating-point support.
-cl /LD /O1 /GS- /nologo ^
-    /I"%JNI_INCLUDE%" /I"%JNI_INCLUDE_WIN32%" ^
-    "%A11Y_SRC%" ^
-    /Fe:"%OUT_DIR_X64%\nucleus_tao_a11y.dll" ^
-    /link /NODEFAULTLIB /ENTRY:DllMain ^
-    kernel32.lib user32.lib comctl32.lib ole32.lib oleaut32.lib uuid.lib uiautomationcore.lib
-if errorlevel 1 (
-    echo ERROR: x64 a11y compilation failed >&2
-    exit /b 1
-)
-
 cl /LD /O1 /GS- /nologo ^
     /I"%JNI_INCLUDE%" /I"%JNI_INCLUDE_WIN32%" ^
     "%DND_SRC%" ^
@@ -231,17 +216,6 @@ if errorlevel 1 (
     goto :clear_cache
 )
 
-cl /LD /O1 /GS- /nologo ^
-    /I"%JNI_INCLUDE%" /I"%JNI_INCLUDE_WIN32%" ^
-    "%A11Y_SRC%" ^
-    /Fe:"%OUT_DIR_ARM64%\nucleus_tao_a11y.dll" ^
-    /link /ENTRY:DllMain ^
-    kernel32.lib user32.lib comctl32.lib ole32.lib oleaut32.lib uuid.lib uiautomationcore.lib
-if errorlevel 1 (
-    echo WARNING: ARM64 a11y compilation failed >&2
-    endlocal
-    goto :clear_cache
-)
 
 cl /LD /O1 /GS- /nologo ^
     /I"%JNI_INCLUDE%" /I"%JNI_INCLUDE_WIN32%" ^
